requirements.txt
datasets==4.1.1
faiss==1.9.0
gymnasium==1.2.1
langchain_community==0.4.1
matplotlib==3.10.7
networkx==3.5
numpy==2.3.4
openai==2.6.1
pandas==1.5.3
plotly==6.3.1
psutil==7.1.0
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence_transformers==5.1.1
torch==2.6.0+cu124
wordcloud==1.9.4
